src/main/java/me/nolanjames/theoldtreeapi/category/CategoryController.java
package me.nolanjames.theoldtreeapi.category;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import me.nolanjames.theoldtreeapi.category.dto.CategoryRequest;
import me.nolanjames.theoldtreeapi.category.dto.CategoryResponse;
import me.nolanjames.theoldtreeapi.category.service.CategoryService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/category")
@RequiredArgsConstructor
public class CategoryController {

    private final CategoryService categoryService;

    @PostMapping()
    public ResponseEntity<CategoryResponse> createCategory(@Valid @RequestBody CategoryRequest categoryRequest) {
        return new ResponseEntity<>(categoryService.createCategory(categoryRequest), HttpStatus.CREATED);
    }

    @GetMapping()
    public ResponseEntity<List<CategoryResponse>> listCategories() {
        return ResponseEntity.ok(categoryService.listCategories());
    }

    @GetMapping({"name"})
    public ResponseEntity<CategoryResponse> getCategoryByName(@RequestParam(name = "name") String name) {
        return ResponseEntity.ok(categoryService.findByName(name));
    }
}
